Add functions for checking if either player won
Before this change, no one won or tied the game. Now, given a
randomly generated tic-tac-toe board, a function can determine if
either the player, or the opponent has won the game in any way
possible. This function also generates a message indicating who won.

With a randomly generated board, this function is biased for P1
because the conditional checks P1 first and every value is either
an X for the player, or an O for the opponent. If the board were
checked each turn, this function wouldn't have that effect and it
would work properly.