Add fns for preventing errors and fix winner fn
Before the game would allow an invalid string for the coordinates
of the token. The game would also allow a token to be placed on a
location that was already taken by a player.

The checkWhoWon() function gave the right response to the player,
but not the computer.

Lastly, a tie function was added in order
to announce a tie if the entire board is occupied by tokens.