Add ability to place tokens via DOM manipulation
The game is playable through the console or using the webpage.
In order to accomplish this, first, the prompts to the user had
to be removed. Any code that manipulated an expected string had
to be changed to expect two numbers.

Then, each square upon creation was given a row and column property
denoting its location on the board as it was created from the board
array. Then the playRound() function was able to use those numbers.

In addition, the board waits for a response after a player has scored
a point before restarting the game.

Lastly, the restart button clears the board.